move bit helper into join.py, improve test readability
