+++ python/cuspatial/cuspatial/tests/spatial/join/test_point_in_polygon.py
@@ -5,7 +5,7 @@ import numpy as np
 import cudf
 
 import cuspatial
-from cuspatial.utils import gis_utils
+from cuspatial.core.spatial.join import pip_bitmap_column_to_binary_array
 
 
 def test_empty():
@@ -228,25 +228,31 @@ def test_three_points_two_features():
 
 
 def test_pip_bitmap_column_to_binary_array():
-    col = cudf.Series([0, 13, 3, 9])._column
-    got = gis_utils.pip_bitmap_column_to_binary_array(col, width=4)
+    col = cudf.Series([0b00000000, 0b00001101, 0b00000011, 0b00001001])._column
+    got = pip_bitmap_column_to_binary_array(col, width=4)
     expected = np.array(
         [[0, 0, 0, 0], [1, 1, 0, 1], [0, 0, 1, 1], [1, 0, 0, 1]], dtype="int8"
     )
     np.testing.assert_array_equal(got.copy_to_host(), expected)
 
     col = cudf.Series([])._column
-    got = gis_utils.pip_bitmap_column_to_binary_array(col, width=0)
+    got = pip_bitmap_column_to_binary_array(col, width=0)
     expected = np.array([], dtype="int8").reshape(0, 0)
     np.testing.assert_array_equal(got.copy_to_host(), expected)
 
     col = cudf.Series([None, None], dtype="float64")._column
-    got = gis_utils.pip_bitmap_column_to_binary_array(col, width=0)
+    got = pip_bitmap_column_to_binary_array(col, width=0)
     expected = np.array([], dtype="int8").reshape(2, 0)
     np.testing.assert_array_equal(got.copy_to_host(), expected)
 
-    col = cudf.Series([238, 13, 29594])._column
-    got = gis_utils.pip_bitmap_column_to_binary_array(col, width=15)
+    col = cudf.Series(
+        [
+            0b000000011101110,
+            0b000000000001101,
+            0b111001110011010,
+        ]
+    )._column
+    got = pip_bitmap_column_to_binary_array(col, width=15)
     expected = np.array(
         [
             [0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 0],
@@ -258,6 +264,6 @@ def test_pip_bitmap_column_to_binary_array():
     np.testing.assert_array_equal(got.copy_to_host(), expected)
 
     col = cudf.Series([0, 0, 0])._column
-    got = gis_utils.pip_bitmap_column_to_binary_array(col, width=3)
+    got = pip_bitmap_column_to_binary_array(col, width=3)
     expected = np.array([[0, 0, 0], [0, 0, 0], [0, 0, 0]], dtype="int8")
     np.testing.assert_array_equal(got.copy_to_host(), expected)
